refactor: improve readability
